+++ Sources/Sentry/SentrySessionReplayIntegration.m
@@ -5,14 +5,18 @@
 #    import "SentryClient+Private.h"
 #    import "SentryDependencyContainer.h"
 #    import "SentryDisplayLinkWrapper.h"
+#    import "SentryEvent+Private.h"
 #    import "SentryFileManager.h"
 #    import "SentryGlobalEventProcessor.h"
 #    import "SentryHub+Private.h"
+#    import "SentryLog.h"
 #    import "SentryNSNotificationCenterWrapper.h"
 #    import "SentryOptions.h"
 #    import "SentryRandom.h"
 #    import "SentrySDK+Private.h"
 #    import "SentryScope+Private.h"
+#    import "SentrySerialization.h"
+#    import "SentrySessionReplaySyncC.h"
 #    import "SentrySwift.h"
 #    import "SentrySwizzle.h"
 #    import "SentryUIApplication.h"
@@ -38,6 +42,7 @@ SentrySessionReplayIntegration ()
     BOOL _startedAsFullSession;
     SentryReplayOptions *_replayOptions;
     SentryNSNotificationCenterWrapper *_notificationCenter;
+    SentryOnDemandReplay *_resumeReplayMaker;
 }
 
 - (BOOL)installWithOptions:(nonnull SentryOptions *)options
@@ -61,12 +66,107 @@ SentrySessionReplayIntegration ()
 
     [SentryGlobalEventProcessor.shared
         addEventProcessor:^SentryEvent *_Nullable(SentryEvent *_Nonnull event) {
-            [self.sessionReplay captureReplayForEvent:event];
-
+            if (event.isCrashEvent) {
+                [self resumePreviousSessionReplay:event];
+            } else {
+                [self.sessionReplay captureReplayForEvent:event];
+            }
             return event;
         }];
 
     return YES;
+}
+
+- (void)resumePreviousSessionReplay:(SentryEvent *)event
+{
+    NSURL *dir = [self replayDirectory];
+    NSData *lastReplay =
+        [NSData dataWithContentsOfURL:[dir URLByAppendingPathComponent:@"lastreplay"]];
+    if (lastReplay == nil) {
+        return;
+    }
+
+    NSDictionary<NSString *, id> *jsonObject = [NSJSONSerialization JSONObjectWithData:lastReplay
+                                                                               options:0
+                                                                                 error:nil];
+    SentryId *replayId = jsonObject[@"replayId"]
+        ? [[SentryId alloc] initWithUUIDString:jsonObject[@"replayId"]]
+        : [[SentryId alloc] init];
+    NSURL *lastReplayURL = [dir URLByAppendingPathComponent:jsonObject[@"path"]];
+
+    SentryCrashReplay crashInfo = { 0 };
+    bool hasCrashInfo = sentrySessionReplaySync_readInfo(&crashInfo,
+        [[lastReplayURL URLByAppendingPathComponent:@"crashInfo"].path
+            cStringUsingEncoding:NSUTF8StringEncoding]);
+
+    SentryReplayType type = hasCrashInfo ? SentryReplayTypeSession : SentryReplayTypeBuffer;
+    NSTimeInterval duration
+        = hasCrashInfo ? _replayOptions.sessionSegmentDuration : _replayOptions.errorReplayDuration;
+    int segmentId = hasCrashInfo ? crashInfo.segmentId + 1 : 0;
+
+    _resumeReplayMaker = [[SentryOnDemandReplay alloc] initWithContentFrom:lastReplayURL.path];
+    _resumeReplayMaker.bitRate = _replayOptions.replayBitRate;
+    _resumeReplayMaker.videoScale = _replayOptions.sizeScale;
+
+    NSDate *beginning = hasCrashInfo
+        ? [NSDate dateWithTimeIntervalSinceReferenceDate:crashInfo.lastSegmentEnd]
+        : [_resumeReplayMaker oldestFrameDate];
+    if (beginning == nil) {
+        // no frames to send
+        return;
+    }
+
+    NSError *replayError = nil;
+
+    [_resumeReplayMaker
+        createVideoWithBeginning:beginning
+                             end:[beginning dateByAddingTimeInterval:duration]
+                   outputFileURL:[lastReplayURL URLByAppendingPathComponent:@"lastVideo.mp4"]
+                           error:&replayError
+                      completion:^(SentryVideoInfo *video, NSError *error) {
+                          if (error != nil) {
+                              SENTRY_LOG_ERROR(@"Could not create replay video: %@", error);
+                          } else {
+                              [self captureVideo:video
+                                        replayId:replayId
+                                       segmentId:segmentId
+                                            type:type];
+                          }
+                          self->_resumeReplayMaker = nil;
+                      }];
+
+    if (replayError != nil) {
+        SENTRY_LOG_ERROR(@"Could not create replay video: %@", replayError);
+    } else {
+        NSMutableDictionary *eventContext = event.context.mutableCopy;
+        eventContext[@"replay"] = @{ @"replay_id" : replayId.sentryIdString };
+        event.context = eventContext;
+    }
+}
+
+- (void)captureVideo:(SentryVideoInfo *)video
+            replayId:(SentryId *)replayId
+           segmentId:(int)segment
+                type:(SentryReplayType)type
+{
+    SentryReplayEvent *replayEvent = [[SentryReplayEvent alloc] initWithEventId:replayId
+                                                           replayStartTimestamp:video.start
+                                                                     replayType:type
+                                                                      segmentId:segment];
+    replayEvent.timestamp = video.end;
+    SentryReplayRecording *recording = [[SentryReplayRecording alloc] initWithSegmentId:segment
+                                                                                  video:video
+                                                                            extraEvents:@[]];
+
+    [SentrySDK.currentHub captureReplayEvent:replayEvent
+                             replayRecording:recording
+                                       video:video.path];
+
+    NSError *error = nil;
+    if (![[NSFileManager defaultManager] removeItemAtURL:video.path error:&error]) {
+        NSLog(@"[SentrySessionReplay:%d] Could not delete replay segment from disk: %@", __LINE__,
+            error.localizedDescription);
+    }
 }
 
 - (void)startSession
@@ -110,9 +210,7 @@ SentrySessionReplayIntegration ()
      breadcrumbConverter:(id<SentryReplayBreadcrumbConverter>)breadcrumbConverter
              fullSession:(BOOL)shouldReplayFullSession
 {
-    NSURL *docs =
-        [NSURL fileURLWithPath:[SentryDependencyContainer.sharedInstance.fileManager sentryPath]];
-    docs = [docs URLByAppendingPathComponent:SENTRY_REPLAY_FOLDER];
+    NSURL *docs = [self replayDirectory];
     NSString *currentSession = [NSUUID UUID].UUIDString;
     docs = [docs URLByAppendingPathComponent:currentSession];
 
@@ -125,6 +223,7 @@ SentrySessionReplayIntegration ()
 
     SentryOnDemandReplay *replayMaker = [[SentryOnDemandReplay alloc] initWithOutputPath:docs.path];
     replayMaker.bitRate = replayOptions.replayBitRate;
+    replayMaker.videoScale = replayOptions.sizeScale;
     replayMaker.cacheMaxSize
         = (NSInteger)(shouldReplayFullSession ? replayOptions.sessionSegmentDuration + 1
                                               : replayOptions.errorReplayDuration + 1);
@@ -153,6 +252,39 @@ SentrySessionReplayIntegration ()
                             selector:@selector(resume)
                                 name:UIApplicationWillEnterForegroundNotification
                               object:nil];
+
+    [self saveCurrentSessionInfo:self.sessionReplay.sessionReplayId
+                            path:docs.path
+                         options:replayOptions];
+}
+
+- (NSURL *)replayDirectory
+{
+    NSURL *dir =
+        [NSURL fileURLWithPath:[SentryDependencyContainer.sharedInstance.fileManager sentryPath]];
+    return [dir URLByAppendingPathComponent:SENTRY_REPLAY_FOLDER];
+}
+
+- (void)saveCurrentSessionInfo:(SentryId *)sessionId
+                          path:(NSString *)path
+                       options:(SentryReplayOptions *)options
+{
+    NSDictionary *info = @{
+        @"replayId" : sessionId.sentryIdString,
+        @"path" : path.lastPathComponent,
+        @"errorSampleRate" : @(options.errorSampleRate)
+    };
+    NSData *data = [SentrySerialization dataWithJSONObject:info];
+
+    NSString *infoPath =
+        [[path stringByDeletingLastPathComponent] stringByAppendingPathComponent:@"lastreplay"];
+    if ([NSFileManager.defaultManager fileExistsAtPath:infoPath]) {
+        [NSFileManager.defaultManager removeItemAtPath:infoPath error:nil];
+    }
+    [data writeToFile:infoPath atomically:YES];
+
+    sentrySessionReplaySync_start([[path stringByAppendingPathComponent:@"crashInfo"]
+        cStringUsingEncoding:NSUTF8StringEncoding]);
 }
 
 - (void)stop
@@ -185,9 +317,9 @@ SentrySessionReplayIntegration ()
     [self startSession];
 }
 
-- (void)captureReplay
+- (BOOL)captureReplay
 {
-    [self.sessionReplay captureReplay];
+    return [self.sessionReplay captureReplay];
 }
 
 - (void)configureReplayWith:(nullable id<SentryReplayBreadcrumbConverter>)breadcrumbConverter
@@ -291,6 +423,9 @@ SentrySessionReplayIntegration ()
     [SentrySDK.currentHub captureReplayEvent:replayEvent
                              replayRecording:replayRecording
                                        video:videoUrl];
+
+    sentrySessionReplaySync_updateInfo(
+        (unsigned int)replayEvent.segmentId, replayEvent.timestamp.timeIntervalSinceReferenceDate);
 }
 
 - (void)sessionReplayStartedWithReplayId:(SentryId *)replayId